wg_control/requirements.txt
pytz
requests

djangorestframework==3.13.1
django-extensions==3.1.0

Django==3.2.13
django-timezone-field 
django-cors-headers
whitenoise

psycopg2-binary==2.8.6
pusher==3.0.0
python-decouple==3.4

celery== 5.2.6
django-celery-beat==2.2.1
daphne
uvloop
httptools

redis==4.2.2
dj-database-url

django-inlinecss
pyswagger
qrcode
image


